new: Features order CRUD

+++ src/modules/database/migrations/20210328232826_order.ts
@@ -4,7 +4,7 @@ export async function up(knex: Knex): Promise<void> {
   await knex.schema.createTable('Order', table => {
     table.increments('id').primary();
     table.string('description', 50).notNullable();
-    table.integer('amount').notNullable();
+    table.decimal('amount').notNullable();
     table.decimal('value').notNullable();
     table.dateTime('createdDate').notNullable();
     table.dateTime('updatedDate').notNullable();
